implement basic client functionaloty, rework processor
