+++ core/node/proof_data_handler/Cargo.toml
@@ -24,6 +24,10 @@ axum.workspace = true
 tokio.workspace = true
 tower-http = { workspace = true, features = ["compression-zstd", "decompression-zstd"] }
 tracing.workspace = true
+reqwest = {workspace = true, features = ["rustls-tls"]}
+serde = { workspace = true, features = ["derive"] }
+
+zksync_task_management.workspace = true
 
 [dev-dependencies]
 hyper.workspace = true
