+++ core/node/proof_data_handler/src/tests.rs
@@ -6,7 +6,7 @@ use axum::{
 };
 use serde_json::json;
 use tower::ServiceExt;
-use zksync_config::configs::{ProofDataHandlerConfig, TeeConfig};
+use zksync_config::configs::{proof_data_handler::ApiMode, ProofDataHandlerConfig, TeeConfig};
 use zksync_dal::{ConnectionPool, CoreDal};
 use zksync_object_store::MockObjectStore;
 use zksync_prover_interface::api::SubmitTeeProofRequest;
@@ -32,6 +32,10 @@ async fn request_tee_proof_inputs() {
                 tee_proof_generation_timeout_in_secs: 600,
                 tee_batch_permanently_ignored_timeout_in_hours: 10 * 24,
             },
+            api_mode: ApiMode::default(),
+            gateway_api_url: None,
+            proof_fetch_interval_in_secs: 10,
+            proof_gen_data_submit_interval_in_secs: 10,
         },
         L1BatchCommitmentMode::Rollup,
         L2ChainId::default(),
@@ -92,6 +96,10 @@ async fn submit_tee_proof() {
                 tee_proof_generation_timeout_in_secs: 600,
                 tee_batch_permanently_ignored_timeout_in_hours: 10 * 24,
             },
+            api_mode: ApiMode::default(),
+            gateway_api_url: None,
+            proof_fetch_interval_in_secs: 10,
+            proof_gen_data_submit_interval_in_secs: 10,
         },
         L1BatchCommitmentMode::Rollup,
         L2ChainId::default(),
