+++ prover/crates/bin/prover_fri_gateway/src/main.rs
@@ -2,18 +2,18 @@ use std::time::Duration;
 
 use anyhow::Context as _;
 use clap::Parser;
-use proof_gen_data_fetcher::ProofGenDataFetcher;
-use proof_submitter::ProofSubmitter;
+use client::proof_gen_data_fetcher::ProofGenDataFetcher;
+use client::proof_submitter::ProofSubmitter;
+use server::Processor;
 use tokio::sync::{oneshot, watch};
 use traits::PeriodicApi as _;
+use zksync_config::configs::proof_data_handler::ApiMode;
 use zksync_core_leftovers::temp_config_store::{load_database_secrets, load_general_config};
 use zksync_env_config::object_store::ProverObjectStoreConfig;
 use zksync_object_store::ObjectStoreFactory;
 use zksync_prover_dal::{ConnectionPool, Prover};
 use zksync_task_management::ManagedTasks;
 use zksync_vlog::prometheus::PrometheusExporterConfig;
-use client::proof_gen_data_fetcher::ProofGenDataFetcher;
-use client::proof_submitter::ProofSubmitter;
 
 mod client;
 mod metrics;
@@ -26,6 +26,10 @@ async fn main() -> anyhow::Result<()> {
 
     let general_config = load_general_config(opt.config_path).context("general config")?;
     let database_secrets = load_database_secrets(opt.secrets_path).context("database secrets")?;
+
+    let data_handler_config = general_config.proof_data_handler_config.context("data handler config")?;
+
+    let api_mode = data_handler_config.api_mode;
 
     let observability_config = general_config
         .observability
@@ -53,17 +57,6 @@ async fn main() -> anyhow::Result<()> {
     );
     let store_factory = ObjectStoreFactory::new(object_store_config.0);
 
-    let proof_submitter = ProofSubmitter::new(
-        store_factory.create_store().await?,
-        config.api_url.clone(),
-        pool.clone(),
-    );
-    let proof_gen_data_fetcher = ProofGenDataFetcher::new(
-        store_factory.create_store().await?,
-        config.api_url.clone(),
-        pool,
-    );
-
     let (stop_sender, stop_receiver) = watch::channel(false);
 
     let (stop_signal_sender, stop_signal_receiver) = oneshot::channel();
@@ -75,26 +68,59 @@ async fn main() -> anyhow::Result<()> {
     })
     .context("Error setting Ctrl+C handler")?;
 
-    tracing::info!("Starting Fri Prover Gateway");
+    tracing::info!("Starting Fri Prover Gateway in mode {:?}", api_mode);
 
-    let tasks = vec![
-        tokio::spawn(
-            PrometheusExporterConfig::pull(config.prometheus_listener_port)
-                .run(stop_receiver.clone()),
-        ),
-        tokio::spawn(proof_gen_data_fetcher.run(config.api_poll_duration(), stop_receiver.clone())),
-        tokio::spawn(proof_submitter.run(config.api_poll_duration(), stop_receiver)),
-    ];
+    let tasks = match api_mode {
+        ApiMode::Legacy => {
+            let proof_submitter = ProofSubmitter::new(
+                store_factory.create_store().await?,
+                config.api_url.clone(),
+                pool.clone(),
+            );
+            let proof_gen_data_fetcher = ProofGenDataFetcher::new(
+                store_factory.create_store().await?,
+                config.api_url.clone(),
+                pool,
+            );
+
+            vec![
+                tokio::spawn(
+                    PrometheusExporterConfig::pull(config.prometheus_listener_port)
+                        .run(stop_receiver.clone()),
+                ),
+                tokio::spawn(proof_gen_data_fetcher.run(config.api_poll_duration(), stop_receiver.clone())),
+                tokio::spawn(proof_submitter.run(config.api_poll_duration(), stop_receiver)),
+            ]
+        }
+        ApiMode::ProverCluster => {
+            let processor = Processor::new(
+                store_factory.create_store().await?,
+                pool,
+            );
+
+            let api = server::Api::new(processor.clone(), config.port);
+
+            vec![
+                tokio::spawn(
+                    PrometheusExporterConfig::pull(config.prometheus_listener_port)
+                        .run(stop_receiver.clone()),
+                ),
+                tokio::spawn(api.run(stop_receiver)),
+            ]
+        }
+    };
 
     let mut tasks = ManagedTasks::new(tasks);
     tokio::select! {
         _ = tasks.wait_single() => {},
-        _ = stop_signal_receiver => {
+         _ = stop_signal_receiver => {
             tracing::info!("Stop signal received, shutting down");
         }
     }
+            
     stop_sender.send(true).ok();
     tasks.complete(Duration::from_secs(5)).await;
+
     Ok(())
 }
